utils/github: split module
